ARM: DaVinci: support two gpio expanders
Declare two of the three I2C GPIO expanders to the EVM board init logic, and
use them.  One hooks up to the LEDs using the leds-gpio driver; the other
exposes a switch to sysfs, and initializes the audio clocks.

The third expander is currently handled by gpio_expander_davinci.c (which
has not yet gone upstream, and probably can't), so a patch to handle that
expander should remove that file and its users.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@mvista.com>