ARM: DaVinci: Standardise deceased cows to fix MAC address
The I2C EMAC code for reading the MAC address from EEPROM uses
deadbeEf in one place and deadbeAf in others as a magic string. As a
result, all DVEVMs actually get the same default MAC address. Here,
we correct them to the British English spelling of beef to fix the
issue and avoid further confusion (hopefully).

Signed-off-by: Bernard Blackham <bernard@largestprime.net>
Signed-off-by: Kevin Hilman <khilman@mvista.com>