ARM: DaVinci: remove legacy GPIO expanders in favor of new GPIO framework
Remove the two (!) legacy GPIO expander modules for DaVinci U35,
and provide standard GPIO framework support for that chip.  It's
initialized so that IDE drives work.  The EVM support in the musb
driver now uses GPIO calls to access that signal.

This breaks the build for the RTC driver.  Probably all the msp430
related code should get updated to use the MFD framework.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>