dm355evm input driver
Simple input driver support for the events reported by the
MSP430 firmware on the DM355 EVM.  Verified using the remote
included with the kit; docs weren't quite right.  Some of
the keycode selections might need improvement; I'm hoping
I implemented the remapping support right, so there's at
least a runtime workaround.

Keys on the remote seem to repeat undesirably.  Someone might
want to see what's up with that, and fix it; it might just
be an issue of tracking RC5 state and using the toggle.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>