mmc device creation updates (MMC1 too)
Update MMC/SD device initialization to handle more chips, in
particular the DM355 with its second MMC/SD controller:

 - Chip-specific stuff:
    * dm646x chips have no MMC
    * if this is a dm355, declare the second controller

 - Completeness/correctness:
    * use new davinci_clk_associate(), clockname "mmc"
    * provide DMA channel resources
    * provide SDIO IRQ resources
    * device numbering starts from zero, not one

 - Claim enough I/O space for all mmc/sd registers ...
   in fact, the whole page that's being ioremapped!

Note that this doesn't address the basic problem that there's no way
now for board init code to say which controllers to initialize, or
how to do so.  Seems to not be an issue for any of the boards currently
in the DaVinci GIT tree.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>