mmc platform_data for dm6446 evm
Update DM6466 EVM to support the new MMC board config data.

This provides a miniature I2C driver in the board-dm644x-evm.c
file ... unusual, but it does make it easier to pass the key
information easily to the MMC driver.

There's no attempt to use the gpio_to_irq(GPIO(7)) signals that
are issued when the various card detect signals change, or when
the IR remote sends a command.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>