davinci_nand separate dm355evm and dm6446evm specifics
Start splitting out the board-specific NAND support better,
by having board-specific flash init functions.  DM355 EVM
has a newer and faster chip.

It's still in the wrong place, but at least it's better
factored now ... which will help later when it's moved to
the board-*.c files.  (NOTE:  sffsdr support is missing;
it might not actually need to set the EMIF A1CR register.)

Also update the header comment slightly, listing key
assumptions made by this driver ... and delete a broken
comment in the dm6446 setup code, that's not the value
it was writing.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>