usb fixes for dm6446evm
Fix an init sequence bug with the MUSB driver on a DaVinci EVM:
don't initialize it until after the pcf8574 GPIO expander is ready.
This prevent oopsing.

And since the DaVinci GPIOs have been renumbered ... use the new number
for the GPIO on that pcf8574 chip.  (Eventually that number should be
passed through platform_data to the driver.)

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>