dm355: mmc0 works for pio
A "cat /proc/iomem" showed why mmc0 is now broken on dm355:

  01e00000-01e00fff : davinci_mmc.1
    01e00000-01e00fff : davinci_mmc
  01e10000-01e10fff : davinci_mmc.0
    01e10000-01e10fff : davinci_nand
      01e10000-01e10fff : davinci_mmc
  02000000-03ffffff : davinci_nand

What was the MMC0 controller address on dm6446 chips is the AEMIF
control register bank on dm355 chips, used by the NAND driver.

Use the right address, and now MMC0 works.  (In PIO mode, and
with that ASM-to-C conversion patch reverted.)  And for that
matter, the dm9000 starts working again.  Win!  Maybe the NAND
driver will be happier too...

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>