SD/MMC high speed support
In place of a FIXME, enable high speed SD (and MMC) support.
I've seen comfortably more than 6 MByte/sec with "hdparm -tT".

Plus minor fixes:  cleanly fall back to PIO if for some reason
DMA init fails; report the *actual* mode (DMA/PIO) not the one
that was attempted; use more conventional labels in /proc/iomem;
and remove an end-of-line whitespace.

Note that on the DM355 EVM cards won't clock over 27 MHz since the
functional clock is at 108 MHz, and the divider doesn't allow a
divide-by-three.  Hmm, the MMC controller specs say that it only
supports up to 100 MHz for a functional clock...

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>