EDMA renames: all remaining operations
This patch primarily renames functions to use an edma_ prefix and remove
a needless infix "_dma_" token.  Again, parameters which identify parameter
RAM slots were renamed as "slot" from "lch".

Also, edma_get_position() was moved to be adjacent to the other operations
that rely on just a few fields of a parameter RAM slot.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>