ARM DaVinci: Low level debug macros for DaVinci Platforms
This patch defines debug macros for low-level debugging for Davinci based
platforms

Tested on :
        - DM644x DaVinci EVM
        - DM646X DaVinciHD EVM
	- DM355 EVM

This patch attempts to solve the low-level debug issue in DM646x. The UART
on DM646x SoC allows only 32-bit access. The existing debug-macro.S uses the
macros from debug-8250.S file. This led to garbage serial out in the case of
DM646x.

The inclusion of debug-8250.S does not allow for run time fix for this issue.
There are compile time errors due to multiple definitions of the macros.
Also when building a single image for multiple DaVinci Platforms, the ifdefs
cannot be relied upon.

The solution below does not include the debug-8250.S file and defines the
necessary macros. This solution was arrived at after observing that word
access does not affect the low-level debug messages on DM644x/DM355.

The other approach to this issue is to use the UART module information
available in the peripheral registers to decide the access mechanism. But this
will have to be done for every access of UART specifically for DM646x. Also
this calls for a modification of the debug-8250.S file.

Signed-off-by: Chaithrika U S <chaithrika@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>