davinci_nand: clean up the A1CR hacks
The A1CR hacks aren't really needed on the DM355 EVM; NAND timings
set by UBL (or whatever) are a bit slower than needed, but that's
not too bad.  Remove, but add a comment in board setup.

For the DM6446 EVM, the problem is how to cope with re-jumpering
in the default boot-from-NOR case.  Boards modified to boot from
NAND should already have A1CR timings set up.  Comment this; for
the moment it's a bit nasty to move this code to board setup.
(And remove pointless AWRCC setting.  The "Extended Wait" mode
doesn't work with NAND chips.)

Related -- make board-dm6446-evm.c warn about trying to make NAND
and NOR coexist, and only set up one of them.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>