more clock/mux splitting: ASP/ASoC
Decouple ASP clk_*() calls from pinmmux for DM355 and DM6446.
This removes the need for dm355_psc_mux(), and is a net minor
source and runtime code shrink.

Note that this keeps the ASoC-related board setup code together,
but unfortunately that still doesn't live in mach-davinci with
all the other board setup logic.  On my DM6446 EVM this muxing
is needed; u-boot doesn't.  On the DM355 EVM, it isn't needed;
u-boot sets that up.  I have no idea if SFFSDR boards need it.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>