dm355 generic spi0 setup
Init the DM355 SPI0 device.  Board init code would call this, which
would handle all pinmuxing needed.

This configures the second SPI interrupt for this controller; it's
not clear we'll actually want to use those controller IRQs ... but
the just-posted driver expects to have one.  (Pure PIO doesn't need
them, it can poll; and DMA can be driven by EDMA irqs.)  But it does
seem that at least on dm355 we'd want to avoid the first IRQ because
of irq mux conflicts with EDMA.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>