MMC: davinci: remove asm code and multiple of 4-byte restriction
This patch will remove asm code which has been used in the PIO mode.
After this patch, it will support write/read request which is not mutiple
of 4 bytes and in order to support this both writel/readl and
iowrite8_rep/ioread8_rep are used. It has been tested in DMA and PIO mode on
the DaVinci platform.

Signed-off-by: Purushotam Kumar <purushotam@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>