Implement the new memory_accessor interface for I2C EEPROM
In the case of at24, the platform code registers a 'setup' callback
with the at24_platform_data.  When the at24 driver detects an EEPROM,
it fills out the read and write functions of the memory_accessor and
calls the setup callback passing the memory_accessor struct.  The
platform code can then use the read/write functions in the
memory_accessor struct for reading and writing the EEPROM.

Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>