ARM: DaVinci: EMAC: RMII bit manipulation
This patch is useful for OMAP-L137 based boards which may have RMII enabled.

EMAC on OMAP-L137 supports RMII and on enabling this the RMIISPEED MASK has to
be set/cleared for 100/10 Mbps operation respectively. A member has been added
to the EMAC platform data to indicate the type of interface.

Signed-off-by: Chaithrika U S <chaithrika@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>