ARM: DaVinci: EMAC: Remove cpu_is_xxx() macros
Remove the usage of cpu_is_xxx() macros from the EMAC driver.

Add a member to the platform data to indicate the version of the EMAC.
Use this information instead of the cpu_is_xxx() macros.

This patch is dependent on the previous sets of patches submitted for EMAC.
The changes have been tested on TI DM644x and DM646x EVMs.

Signed-off-by: Chaithrika U S <chaithrika@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>