ARM: DaVinci: Clear an EDMA event
Adds edma_clear_event API support to EDMA.

There has been a change to the MMC interface on TI's DA830/OMAP-L137
architecture that requires setting the DMATRIG bit to trigger
the first read data transfer when doing a read operation using
the CPU, but it has the unintended consequence causing a DMA
event.  Because of this, the next DMA read from the MMC/SD will
fail with an Event Miss. This API will be used to clear the ER.

Signed-off-by: David Griego <dgriego@mvista.com>
Signed-off-by: Sergei Shtylyov <sshtylyov@ru.mvista.com>
Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>
Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>