ARM: DaVinci: Fix for MMC/SD PIO mode on DA830/OMAP-L137
Fixes the PIO mode for MMC/SD on TI's DA830/OMAP-L137 architecture.

On TI's DA830/OMAP-L137, the DMATRIG bit should be set for PIO
transfers.  If MMC is being used for both PIO and EDMA based
transfers, then it is likely that, an extra MMC EVENT is latched
in the EDMA Event register even when just the CPU and MMC are in
play. To properly switch from PIO to EDMA, clear any extra or
unexpected event latched in the Event register, by doing a write
to the corresponding bits in the ECR register.

Signed-off-by: Sudhakar Rajashekhara <sudhakar.raj@ti.com>